Integer conversions

+++ acp_times.py
@@ -45,7 +45,7 @@ def open_time( control_dist_km, brevet_dist_km, brevet_start_time ):
     print("control_start_time= " + str(control_start_time))
     minute, hour = math.modf(control_start_time)
     print("hour= " + str(hour))
-    return brevet_start_time.replace(hours=+hour, minutes=+(int(60*minute)))
+    return brevet_start_time.replace(hours=+int(hour), minutes=+(int(60*minute)))
 
 def close_time( control_dist_km, brevet_dist_km, brevet_start_time ):
     """
@@ -78,7 +78,7 @@ def close_time( control_dist_km, brevet_dist_km, brevet_start_time ):
     print("control_close_time= " + str(control_close_time))
     minute, hour = math.modf(control_close_time)
     print("hour= " + str(hour) + " minute= " + str(int(60*minute)))
-    return brevet_start_time.replace(hour=+hour, minute=+(int(60*minute)))
+    return brevet_start_time.replace(hour=+int(hour), minute=+(int(60*minute)))
